fix(#959): move AuditConfig from core/config.py to contracts/types.py
Per KERNEL-ARCHITECTURE, config dataclasses consumed by multiple layers
(factory, services, contracts) should live in the contracts layer.
Move AuditConfig to contracts/types.py and update all 4 importers.
